@@ -13,7 +13,7 @@ namespace SimplCommerce.Module.Contacts
             GlobalConfiguration.RegisterAngularModule("simplAdmin.contacts");
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
         }
     }
