@@ -32,7 +32,7 @@ namespace SimplCommerce.Module.HangfireJobs
             });
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             app.UseHangfire();
             app.InitializeHangfireJobs();
