@@ -17,7 +17,7 @@ namespace SimplCommerce.Module.News
             GlobalConfiguration.RegisterAngularModule("simplAdmin.news");
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
 
         }
