@@ -21,7 +21,7 @@ namespace SimplCommerce.Module.Orders
             GlobalConfiguration.RegisterAngularModule("simplAdmin.orders");
         }
 
-        public void Configure(IApplicationBuilder app, Microsoft.AspNetCore.Hosting.IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, Microsoft.AspNetCore.Hosting.IWebHostEnvironment env)
         {
         }
     }
