@@ -15,7 +15,7 @@ namespace SimplCommerce.Module.SampleData
             services.AddTransient<ISampleDataService, SampleDataService>();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
 
         }
