@@ -13,7 +13,7 @@ namespace SimplCommerce.Module.Search
             GlobalConfiguration.RegisterAngularModule("simplAdmin.search");
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
 
         }
