@@ -17,7 +17,7 @@ namespace SimplCommerce.Module.ShoppingCart
             services.AddTransient<INotificationHandler<UserSignedIn>, UserSignedInHandler>();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
 
         }
