@@ -15,7 +15,7 @@ namespace SimplCommerce.Module.SignalR
             serviceCollection.AddSingleton<IOnlineClientManager, OnlineClientManager>();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             app.UseSignalR(routes =>
             {
