@@ -13,7 +13,7 @@ namespace SimplCommerce.Module.WishList
             serviceCollection.AddTransient<IWishListService, WishListService>();
         }
 
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             
         }
