@@ -82,11 +82,10 @@ namespace SimplCommerce.Module.Catalog.Areas.Catalog.Controllers
                 query = query.Where(x => x.Price <= searchOption.MaxPrice.Value);
             }
 
-            var brands = searchOption.GetBrands();
+            var brands = searchOption.GetBrands().ToArray();
             if (brands.Any())
             {
-                var brandIds = _brandRepository.Query().Where(x => brands.Contains(x.Slug)).Select(x => x.Id).ToList();
-                query = query.Where(x => x.BrandId.HasValue && brandIds.Contains(x.BrandId.Value));
+                query = query.Where(x => x.BrandId != null && brands.Contains(x.Brand.Slug));
             }
 
             model.TotalProduct = query.Count();
@@ -98,16 +97,13 @@ namespace SimplCommerce.Module.Catalog.Areas.Catalog.Controllers
                 offset = (_pageSize * currentPageNum) - _pageSize;
             }
 
-            query = query
-                .Include(x => x.Brand)
-                .Include(x => x.ThumbnailImage);
-
             query = ApplySort(searchOption, query);
 
             var products = query
-                .Select(x => ProductThumbnail.FromProduct(x))
+                .Include(x => x.ThumbnailImage)
                 .Skip(offset)
                 .Take(_pageSize)
+                .Select(x => ProductThumbnail.FromProduct(x))
                 .ToList();
 
             foreach (var product in products)
@@ -145,12 +141,12 @@ namespace SimplCommerce.Module.Catalog.Areas.Catalog.Controllers
             model.FilterOption.Price.MaxPrice = query.Max(x => x.Price);
             model.FilterOption.Price.MinPrice = query.Min(x => x.Price);
 
-            model.FilterOption.Brands = query
-                .Where(x => x.BrandId != null)
+            model.FilterOption.Brands = query.Include(x => x.Brand)
+                .Where(x => x.BrandId != null).ToList()
                 .GroupBy(x => x.Brand)
                 .Select(g => new FilterBrand
                 {
-                    Id = (int)g.Key.Id,
+                    Id = g.Key.Id,
                     Name = g.Key.Name,
                     Slug = g.Key.Slug,
                     Count = g.Count()
