@@ -1,5 +1,5 @@
 ﻿@model ManageLoginsViewModel
-@using Microsoft.AspNetCore.Http.Authentication
+@using Microsoft.AspNetCore.Authentication
 @{
     ViewData["Title"] = "Manage your external logins";
 }
