@@ -14,8 +14,6 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authentication.OAuth;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.CodeAnalysis;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -23,13 +21,11 @@ using Microsoft.IdentityModel.Tokens;
 using Newtonsoft.Json;
 using SimplCommerce.Infrastructure;
 using SimplCommerce.Infrastructure.Modules;
-using SimplCommerce.Infrastructure.Web;
 using SimplCommerce.Infrastructure.Web.ModelBinders;
 using SimplCommerce.Module.Core.Data;
 using SimplCommerce.Module.Core.Extensions;
 using SimplCommerce.Module.Core.Models;
 using Microsoft.AspNetCore.Mvc.ApplicationParts;
-using Microsoft.EntityFrameworkCore.Extensions;
 using Microsoft.Extensions.Localization;
 
 namespace SimplCommerce.WebHost.Extensions
@@ -73,7 +69,6 @@ namespace SimplCommerce.WebHost.Extensions
                 }
 
                 GlobalConfiguration.Modules.Add(module);
-                RegisterModuleInitializerServices(module, ref services);
             }
 
             return services;
@@ -86,23 +81,24 @@ namespace SimplCommerce.WebHost.Extensions
                 {
                     o.EnableEndpointRouting = false;
                     o.ModelBinderProviders.Insert(0, new InvariantDecimalModelBinderProvider());
-                    o.Filters.Add(new AutoValidateAntiforgeryTokenAttribute());
+                    // o.Filters.Add(new AutoValidateAntiforgeryTokenAttribute());
                 })
-                .AddRazorOptions(o =>
-                {
-                    foreach (var module in modules.Where(x => !x.IsBundledWithHost))
-                    {
-                        o.AdditionalCompilationReferences.Add(MetadataReference.CreateFromFile(module.Assembly.Location));
-                    }
-                })
+                //.AddRazorOptions(o =>
+                //{
+                //    foreach (var module in modules.Where(x => !x.IsBundledWithHost))
+                //    {
+                //        o.AdditionalCompilationReferences.Add(MetadataReference.CreateFromFile(module.Assembly.Location));
+                //    }
+                //})
                 .AddViewLocalization()
                 .AddModelBindingMessagesLocalizer(services)
-                .AddDataAnnotationsLocalization(o => {
+                .AddDataAnnotationsLocalization(o =>
+                {
                     var factory = services.BuildServiceProvider().GetService<IStringLocalizerFactory>();
                     var L = factory.Create(null);
-                    o.DataAnnotationLocalizerProvider = (t,f) => L;
-                })                
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                    o.DataAnnotationLocalizerProvider = (t, f) => L;
+                })
+                .AddNewtonsoftJson();
 
             foreach (var module in modules.Where(x => !x.IsBundledWithHost))
             {
@@ -286,16 +282,6 @@ namespace SimplCommerce.WebHost.Extensions
                         module.Assembly = assembly;
                     }
                 }
-            }
-        }
-
-        private static void RegisterModuleInitializerServices(ModuleInfo module, ref IServiceCollection services)
-        {
-            var moduleInitializerType = module.Assembly.GetTypes()
-                    .FirstOrDefault(t => typeof(IModuleInitializer).IsAssignableFrom(t));
-            if ((moduleInitializerType != null) && (moduleInitializerType != typeof(IModuleInitializer)))
-            {
-                services.AddSingleton(typeof(IModuleInitializer), moduleInitializerType);
             }
         }
 
