@@ -1,7 +1,7 @@
-﻿using System;
-using Microsoft.AspNetCore.Hosting;
+﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Serilog;
 using SimplCommerce.Module.Core.Extensions;
@@ -12,23 +12,17 @@ namespace SimplCommerce.WebHost
     {
         public static void Main(string[] args)
         {
-            try
-            {
-                BuildWebHost2(args).Run();
-            }
-            catch(Exception ex)
-            {
-                Console.WriteLine(ex);
-            }
+            BuildWebHost2(args).Build().Run();
         }
 
         // Changed to BuildWebHost2 to make EF don't pickup during design time
-        private static IWebHost BuildWebHost2(string[] args) =>
-            Microsoft.AspNetCore.WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
-                .ConfigureAppConfiguration(SetupConfiguration)
-                .ConfigureLogging(SetupLogging)
-                .Build();
+        private static IHostBuilder BuildWebHost2(string[] args) =>
+            Host.CreateDefaultBuilder(args)
+                .ConfigureWebHostDefaults(webBuilder => {
+                    webBuilder.UseStartup<Startup>();
+                    webBuilder.ConfigureAppConfiguration(SetupConfiguration);
+                    webBuilder.ConfigureLogging(SetupLogging);
+                 });
 
         private static void SetupConfiguration(WebHostBuilderContext hostingContext, IConfigurationBuilder configBuilder)
         {
